Neuro_backend/requirements.txt
# ============================================
# NeuroX backend dependencies — RapidOCR (ONNX Runtime) stack
# Runs on any x86_64 CPU (no AVX requirement).
# Default OCR models are bundled inside rapidocr-onnxruntime wheel.
# ============================================

# Core Web Framework
fastapi>=0.115,<1
uvicorn>=0.34,<1
python-multipart>=0.0.18
python-dotenv>=1.1,<2
pydantic>=2.10,<3

# CORS & HTTP
httpx>=0.28,<1
requests>=2.32,<3

# Database (PostgreSQL)
SQLAlchemy>=2.0,<3
psycopg2-binary>=2.9,<3
asyncpg>=0.30,<1
alembic>=1.14,<2

# OCR — RapidOCR with ONNX Runtime (AVX-free, no paddlepaddle)
rapidocr-onnxruntime>=1.4,<2
onnxruntime>=1.19,<2

# Image Processing
Pillow>=10.3
opencv-python>=4.9,<5
numpy>=1.26,<3

# PDF Processing
PyMuPDF>=1.24,<2
pdf2image>=1.17,<2

# Template filling
python-docx>=1.1,<2

# OpenAI API (for extraction fallback)
openai>=1.50,<2

# Authentication
passlib[bcrypt]>=1.7,<2
bcrypt>=4.0,<5

# Utilities
loguru>=0.7,<1
python-dateutil>=2.9,<3
pydantic-settings>=2.7,<3

# OCR post-processing (also used directly by ocr_filter / ocr_to_schema)
shapely>=2,<3
pyclipper>=1.3,<2

# Development
pytest>=8,<9
pytest-asyncio>=0.25,<1
